Initial plotting, trace_plot works, corner plots works more parameters has been added: The plots is inputted into the plot.py. Dependencies for plot.py has been updated to include Model, LinearSegmentedColormap and Corner.
